move isSameDay from TimeProvider to extension fun
