+++ subprojects/common/time-test-fixtures/src/main/kotlin/com/avito/time/StubTimeProvider.kt
@@ -8,8 +8,6 @@ class StubTimeProvider : TimeProvider {
 
     lateinit var now: Date
 
-    override fun isSameDay(date1: Date, date2: Date): Boolean = timeProvider.isSameDay(date1, date2)
-
     override fun nowInMillis(): Long = timeProvider.nowInMillis()
 
     override fun nowInSeconds(): Long = timeProvider.nowInSeconds()
