+++ subprojects/common/time/src/main/kotlin/com/avito/time/DefaultTimeProvider.kt
@@ -1,6 +1,5 @@
 package com.avito.time
 
-import java.util.Calendar
 import java.util.Date
 import java.util.concurrent.TimeUnit
 
@@ -9,13 +8,6 @@ import java.util.concurrent.TimeUnit
  * consider using https://developer.android.com/studio/write/java8-support#library-desugaring
  */
 class DefaultTimeProvider : TimeProvider {
-
-    override fun isSameDay(date1: Date, date2: Date): Boolean {
-        return isSameDay(
-            Calendar.getInstance().apply { time = date1 },
-            Calendar.getInstance().apply { time = date2 }
-        )
-    }
 
     override fun nowInMillis(): Long = System.currentTimeMillis()
 
@@ -26,11 +18,5 @@ class DefaultTimeProvider : TimeProvider {
     override fun toDate(seconds: Long): Date {
         val millis = TimeUnit.SECONDS.toMillis(seconds)
         return Date(millis)
-    }
-
-    private fun isSameDay(cal1: Calendar, cal2: Calendar): Boolean {
-        return cal1.get(Calendar.ERA) == cal2.get(Calendar.ERA) &&
-            cal1.get(Calendar.YEAR) == cal2.get(Calendar.YEAR) &&
-            cal1.get(Calendar.DAY_OF_YEAR) == cal2.get(Calendar.DAY_OF_YEAR)
     }
 }
