+++ subprojects/common/time/src/main/kotlin/com/avito/time/TimeProvider.kt
@@ -4,8 +4,6 @@ import java.util.Date
 
 interface TimeProvider {
 
-    fun isSameDay(date1: Date, date2: Date): Boolean
-
     fun nowInMillis(): Long
 
     fun nowInSeconds(): Long
