+++ subprojects/gradle/slack/src/main/kotlin/com/avito/slack/TodayMessageCondition.kt
@@ -2,12 +2,13 @@ package com.avito.slack
 
 import com.avito.slack.model.FoundMessage
 import com.avito.time.TimeProvider
+import com.avito.time.isSameDay
 import java.util.Date
 
 class TodayMessageCondition(private val timeProvider: TimeProvider) : SlackMessageUpdateCondition {
 
     override fun updateIf(existingMessage: FoundMessage): Boolean {
-        return timeProvider.isSameDay(slackTimeToDate(existingMessage.timestamp), timeProvider.now())
+        return slackTimeToDate(existingMessage.timestamp).isSameDay(timeProvider.now())
     }
 
     private fun slackTimeToDate(slackTimeStamp: String): Date {
